Added backwards compatibility and options object to CRUD
Added aliases for the old canvas-wrapper
Added aliases for olders versions of this module

Item.get Item.getComplete Item.update Item.delete Items.get Items.getComplete
Items.getOne Items.getOneComplete Items.update Items.delete
All now have an optional query object parameter to be put as the
body/querystring when making the request. In the case of update, the
query object is parsed like the create object and replaces the item's
data in the put request

Also added Item.getFlattened & Items.getFlattened for when you want
the list of all the modulesItems but not the modules. So does the simple
reduce for you. Also has a depth parameter which I need to figure out
how to explain